+++ src/main/java/cn/edu/ustb/sem/order/entity/OrderProcess.java
@@ -66,7 +66,7 @@ public class OrderProcess implements Serializable {
 	@Column(name = "group_name")
 	private Integer group;
 	@Column(name = "measure")
-	private Integer measure;
+	private Double measure;
 	@Column(name = "remark")
 	private String remark;
 	@Column(name = "base")
@@ -133,11 +133,11 @@ public class OrderProcess implements Serializable {
 		this.group = group;
 	}
 
-	public Integer getMeasure() {
+	public Double getMeasure() {
 		return measure;
 	}
 
-	public void setMeasure(Integer measure) {
+	public void setMeasure(Double measure) {
 		this.measure = measure;
 	}
